Add combined search for event and pages
This commit uses features from the pg_search repo that allows for
additional attributes for multisearch. This enables multisearch to make
a filter based on theese attributes. In this commit this is used by
adding a publish_at field to search documents. This is to prevent events
that have yet to be published from being searchable.