Add Erlends changes form earlier
Erlend already had a PR out on the branch fix_event_archive or something
similar. However, the PR have merged with other changes on master,
totalling in 91 files changed in total. I just went through the changed
files and added the changes that were relevant for the archived events
search to this new branch.