Fix archived events search (#220)
* Add Erlends changes form earlier

Erlend already had a PR out on the branch fix_event_archive or something
similar. However, the PR have merged with other changes on master,
totalling in 91 files changed in total. I just went through the changed
files and added the changes that were relevant for the archived events
search to this new branch.

* Remove double pagination

* Add translation for page_gap

* Please rubocop with correct indentation

* Move controller method to event model

* Make rubocop happy

* Increase limit from 228 to 350

* Increase ticket limit from 350 to 400