Give MG-Redaksjonen access to images
As of before this commit MG-Redaksjonen didn't have access to
images. They didn't have access to images from the control panel,
but when creating events (which they can), there was a link to adding an
existing image to the event which wasn't guarded behing an ability check
(like '-if can? :create, Image'), which meant lots of permission denied
errors. The 'upload new image' button was behind such a guard, so they
would never see it. This commit gives the MG-Redaksjonen ability to see,
add existing images to events and upload new images. It also hides the
use existing image button from the events creation form for people that
shouldn't see it. Lastly it translates the messages seen in the events
from when you don't have permissions to see images.