Initial, unfinished, rough onboarding document
I have started writing an onboarding document for new (and old) members
of MG::Web. Samfundet.no is a messy project at times, and it can be
difficult to navigate and understand. This document hopefully tries to
find a middle-way between too much and too little detail in regards to
an introduction to samfundet.no.