Validate publication time on event create
We were not validating publication time when creating new events.
This would result in an ugly error message concerning a PostgreSQL
NotNullViolation immediately after submitting. This is because the
PostgreSQL Search gem (pg_search) creates its own table of which a
column is "publish_at" (which is declared `null: false`). This
"publish_at" column would then interact with the publication_time
column, which would result in a crash.